refactor: unneeded LogItem
