+++ deployer/src/deployment/deploy_layer.rs
@@ -106,7 +106,7 @@ impl From<Log> for persistence::Log {
     }
 }
 
-impl From<Log> for shuttle_common::log::Item {
+impl From<Log> for shuttle_common::LogItem {
     fn from(log: Log) -> Self {
         Self {
             id: log.id,
